Add subtypeUtils
